CareMyPet/Backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
PyJWT==2.9.0
bcrypt==4.2.1
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.32.3
email-validator==2.2.0
pytest==8.3.5
mongomock==4.3.0
# Performance: WSGI servers
gunicorn==21.2.0
waitress==3.0.1
# Performance: response caching
Flask-Caching==2.3.0
# Performance: gzip compression
flask-compress==1.15
